+++ resources/views/bootstrap-4-toast.blade.php
@@ -6,7 +6,7 @@
                 'type' => $flashMessage['type'],
                 'message' => $flashMessage['message'],
                 'title' => $flashMessage['title'],
-                'errorsCount' => $errorsCount,
+                'errorsCount' => $flashErrorsCount,
             ])
         @endforeach
         @foreach ($nowMessages as $nowMessage)
@@ -15,7 +15,7 @@
                 'type' => $nowMessage['type'],
                 'message' => $nowMessage['message'],
                 'title' => $nowMessage['title'],
-                'errorsCount' => $errorsCount,
+                'errorsCount' => $nowErrorsCount,
             ])
         @endforeach
     </div>